- Fix missing admin check when opening secured chests. - Moved database related things to CenotaphDatabase.java
